@@ -19,16 +19,18 @@
     "clsx": "^1.2.1",
     "eslint": "^8.29.0",
     "framer-motion": "^8.4.2",
-    "next": "13.0.7",
+    "next": "^13.2.3",
     "next-auth": "^4.19.2",
     "prisma": "^4.10.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scroll-parallax": "^3.3.2"
   },
   "devDependencies": {
+    "@types/node": "18.14.6",
+    "@types/react": "^18.0.28",
     "autoprefixer": "^10.4.13",
-    "eslint-config-next": "^13.0.7",
+    "eslint-config-next": "^13.2.3",
     "eslint-config-prettier": "^8.5.0",
     "postcss": "^8.4.20",
     "prettier": "2.8.1",
