@@ -1,16 +1,15 @@
-import { useEffect, useMemo, useState } from "react";
+import React, { useEffect, useMemo, useState } from "react";
 import Shape from "./Shape";
 
-const selectRandom = (arr) => arr[Math.floor(Math.random() * arr.length)];
+const selectRandom = (arr: any[]) => arr[Math.floor(Math.random() * arr.length)];
 
-const ShapeRain = ({ count }) => {
-  const [shapesArray, setShapesArray] = useState([]);
+const ShapeRain: React.FC<{ count: number}> = ({ count }) => {
+  const [shapesArray, setShapesArray] = useState<JSX.Element[] | never[]>([]);
 
   const shapes = useMemo(() => {
     const arr = [];
     for (let i = 0; i < count; i++) {
       arr.push(
-        <>
           <Shape
             type={selectRandom(["circle", "triangle", "rect"])}
             id={i}
@@ -24,7 +23,6 @@ const ShapeRain = ({ count }) => {
               opacity: Math.random() * 0.25 + 0.75,
             }}
           />
-        </>,
       );
     }
     return arr;
