@@ -3,21 +3,16 @@ import "../styles/globals.css";
 import NavBar from "../components/NavBar";
 import { SessionProvider } from "next-auth/react";
 import { ParallaxProvider } from "react-scroll-parallax";
+import type { AppProps } from "next/app";
 
-function MyApp({ Component, pageProps: { session, ...pageProps } }) {
+function MyApp({ Component, pageProps: { session, ...pageProps } }: AppProps) {
   return (
     <>
       <SessionProvider session={session}>
         <NavBar />
         <div className="content">
           <ParallaxProvider>
-            {Component.Layout ? (
-              <Component.Layout>
                 <Component {...pageProps} />
-              </Component.Layout>
-            ) : (
-              <Component {...pageProps} />
-            )}
           </ParallaxProvider>
         </div>
       </SessionProvider>
