@@ -1,14 +1,14 @@
 import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import NextAuth from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
+import NextAuth, { NextAuthOptions } from "next-auth";
+import GoogleProvider, { GoogleProfile } from "next-auth/providers/google";
 import prisma from "../../../lib/prisma";
 
-export const authOptions = {
+export const authOptions: NextAuthOptions = {
   adapter: PrismaAdapter(prisma),
   providers: [
     GoogleProvider({
-      clientId: process.env.CLIENT_ID,
-      clientSecret: process.env.CLIENT_SECRET,
+      clientId: process.env.CLIENT_ID!,
+      clientSecret: process.env.CLIENT_SECRET!,
     }),
   ],
   session: {
@@ -17,26 +17,15 @@ export const authOptions = {
   secret: process.env.NEXTAUTH_SECRET,
   callbacks: {
     async signIn({ account, profile }) {
+      if (!account || !profile) return false;
       // future GitHub integration?
       if (account.provider === "google") {
-        return profile.email_verified && profile.email.endsWith("@bxscience.edu");
+        return (profile as GoogleProfile).email_verified && (profile as GoogleProfile).email.endsWith("@bxscience.edu");
       }
       // TODO: actual error handling
       // some form of redirecting back to landing page with a red box
       // saying they used a non bxsci email
       return false;
-    },
-    async jwt({ token, account, profile }) {
-      if (account) {
-        token.accessToken = account.access_token;
-        token.id = profile.sub;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      session.accessToken = token.accessToken;
-      session.user.id = token.sub;
-      return session;
     },
   },
 };
