@@ -1,9 +1,10 @@
 // Returns info of CURRENTLY LOGGED IN user
 // There will probably be routes for getting info of other users
 
+import { NextApiRequest, NextApiResponse } from "next";
 import { getUser } from "../../../lib/server";
 
-export default async function handler(req, res) {
+export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method != "GET") {
     return res.status(405).json({ message: "Method Not Allowed" });
   }
