@@ -1,7 +1,13 @@
+import { Submission } from "@prisma/client";
+import { GetServerSideProps } from "next";
 import Layout from "../../../components/dashboard/Layout";
 import { getSubmission } from "../../../lib/server";
 
-export default function Submission({ submission }) {
+type Props = {
+  submission: Submission;
+};
+
+export const SubmissionPage: React.FC<Props> = ({ submission }) => {
   return (
     <div className="max-w-screen-md p-2 py-6 mx-auto">
       <h1 className="mb-4 text-6xl font-bold text-teal-300">{submission.title}</h1>
@@ -10,9 +16,14 @@ export default function Submission({ submission }) {
   );
 }
 
-Submission.Layout = Layout;
+// SubmissionPage.Layout = Layout;
 
-export async function getServerSideProps({ req, query: { id } }) {
+export const getServerSideProps: GetServerSideProps = async ({ req, query: { id } }) => {
+  if (!id || Array.isArray(id)) {
+    return {
+      notFound: true,
+    };
+  }
   const submission = await getSubmission(req, id);
   if (!submission) {
     return {
@@ -25,4 +36,4 @@ export async function getServerSideProps({ req, query: { id } }) {
       submission,
     },
   };
-}
+};
